protos: move streams to pb
