+++ pkg/pb/gen.go
@@ -1,6 +1,6 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package streams
+package pb
 
 //go:generate protoc --go_out=plugins=grpc:. meta.proto
